Moved ffmpeg stuff to test folder, fixed naming
