+++ AudioToTxt/source/CMakeLists.txt
@@ -43,20 +43,12 @@ option(GGML_VULKAN "Enable Vulkan Support" ON)
 #Vulkan
 find_package(Vulkan REQUIRED)
 
-#FFmpeg
-set(FFMPEG_COMPONENTS avcodec avformat avutil swscale swresample)
-
-find_package(FFmpeg REQUIRED COMPONENTS ${FFMPEG_COMPONENTS})
-
 
 
 target_link_libraries(${PROJECT_NAME} PUBLIC 
-    ${FFMPEG_LIBRARIES}
     Vulkan::Vulkan
     whisper
 )
-target_include_directories(${PROJECT_NAME} PRIVATE ${FFMPEG_INCLUDE_DIRS})
-
 #Create Include Headers
 IF (NOT CMAKE_CURRENT_SOURCE_DIR STREQUAL ${CMAKE_SOURCE_DIR}/AudioToTxt/source)
     file(COPY
