+++ AudioToTxt/tests/test_whisper/main.cpp
@@ -3,29 +3,31 @@
 #include <iostream>
 #include <cstdlib>
 
+#include "audioDecoding.h"
+
 int main() {
    
     
 
-    std::filesystem::path exe_path = std::filesystem::current_path(); // Gets the current working directory (where executable runs)
+    std::filesystem::path exePath = std::filesystem::current_path(); // Gets the current working directory (where executable runs)
 
     // Path to the assets folder relative to the executable's directory
-    std::filesystem::path assets_path =  exe_path / "assets";
-    std::filesystem::path input_video = assets_path / "CS_test.mp4";    
+    std::filesystem::path assetsPath =  exePath / "assets";
+    std::filesystem::path inputVideo = assetsPath / "CS_test.mp4";    
 
-    const std::filesystem::path audio_file = assets_path / "audio.pcm";
-    const std::filesystem::path subtitle_file = assets_path / "subtitles.srt";
-    const std::filesystem::path output_video = assets_path / "output_with_subtitles_turbo.mp4";
-    const std::filesystem::path model_path = exe_path / "models" / "ggml-large-v3-turbo-q5_0.bin";
+    const std::filesystem::path audioFile = assetsPath / "audio.pcm";
+    const std::filesystem::path subtitleFile = assetsPath / "subtitles.srt";
+    const std::filesystem::path outputVideo = assetsPath / "output_with_subtitles_turbo.mp4";
+    const std::filesystem::path modelPath = exePath / "models" / "ggml-base.bin";
 
-    std::cout <<"pwd: " << exe_path << std::endl << "Assests path: " << assets_path << std::endl << "Model path: "<< model_path << std::endl <<  "Input file: " << input_video << std::endl << "subtitle file: " << subtitle_file << std::endl;
-    if (!extractAudio(input_video, audio_file)){
+    std::cout <<"pwd: " << exePath << std::endl << "Assests path: " << assetsPath << std::endl << "Model path: "<< modelPath << std::endl <<  "Input file: " << inputVideo << std::endl << "subtitle file: " << subtitleFile << std::endl;
+    if (!extractAudio(inputVideo, audioFile)){
         std::cout << "Failed to generate formated audio file" << std::endl;
     }
-    if (transcribe_audio(model_path, audio_file, subtitle_file) != 0){
+    if (transcribeAudio(modelPath, audioFile, subtitleFile) != 0){
         std::cout << "Failed to generate subtitles" << std::endl;
     }
-    //if (add_subtitles(input_video, subtitle_file, output_video) != 0) return -1;
+    //if (add_subtitles(inputVideo, subtitleFile, outputVideo) != 0) return -1;
 
     return 0;
 }
